requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-community
python-dotenv
pydantic
wikipedia
duckduckgo-search
ddgs
